Refactor constants file.
