+++ src/constants/colormap.ts
